Extract SSH session management into Pool methods, command formatting into new files
