+++ internal/pool/command_formatter_test.go
